Extract renderPipeline helper
